refactor: reorg Announce to Player
